# api/openapi.yaml
openapi: 3.1.0
info:
  title: Skawld Maintenance API
  version: 0.6.0-phase5
  description: >
    Phase 5 pilot-ready evidence-first maintenance intelligence contract.
    Learned candidates require exact evidence, human review, deterministic
    evaluation, approval authority, and explicit publication. AI quality
    labels are append-oriented. This API exposes no industrial control action
    or permit authority.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/executions:
    get:
      summary: List assigned or authorized executions
      tags: [Executions]
      operationId: listExecutions
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: query, name: site_id, schema: {type: string, format: uuid}}
        - {in: query, name: state, schema: {type: string, enum: [ASSIGNED, IN_PROGRESS, COMPLETED]}}
        - {in: query, name: assigned_to, schema: {type: string, description: "Use me for the current mobile principal."}}
        - {in: query, name: cursor, required: false, schema: {type: string}}
        - {in: query, name: page_size, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
      responses:
        "200":
          description: Executions visible to the principal; field roles are constrained to their own assignments.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor, has_more]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/Execution"}}
                  next_cursor: {type: [string, "null"]}
                  has_more: {type: boolean}
        "401": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
  /api/v1/executions/{executionID}:
    get:
      summary: Get execution detail
      tags: [Executions]
      operationId: getExecution
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/ExecutionID"}]
      responses:
        "200":
          description: Execution with steps and field evidence.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Execution"}
        "404": {$ref: "#/components/responses/Problem"}
  /api/v1/executions/{executionID}/start:
    post:
      summary: Start an execution
      tags: [Executions]
      operationId: startExecution
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ExecutionID"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody: {$ref: "#/components/requestBodies/ExpectedVersion"}
      responses:
        "200": {description: Execution started.}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/executions/{executionID}/steps/{stepID}/complete:
    post:
      summary: Complete an execution step
      tags: [Executions]
      operationId: completeExecutionStep
      description: Returns 409 and a blocked step when a required external prerequisite is absent.
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ExecutionID"}
        - {in: path, name: stepID, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expected_execution_version, expected_step_version]
              properties:
                expected_execution_version: {type: integer, format: int64, minimum: 1}
                expected_step_version: {type: integer, format: int64, minimum: 1}
      responses:
        "200": {description: Step completed.}
        "409":
          description: Version conflict or safety prerequisite blocked the step.
  /api/v1/executions/{executionID}/prerequisite-verifications:
    post:
      summary: Record external prerequisite evidence
      tags: [Executions]
      operationId: recordExternalPrerequisiteVerification
      description: Records evidence from an external PTW/LOTO authority; Skawld is not that authority.
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ExecutionID"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PrerequisiteVerificationCommand"}
      responses:
        "201": {description: External verification evidence recorded.}
        "403": {$ref: "#/components/responses/Problem"}
  /api/v1/executions/{executionID}/measurements:
    post:
      summary: Record a maintenance measurement
      tags: [Executions]
      operationId: recordMeasurement
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ExecutionID"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MeasurementCommand"}
      responses:
        "201":
          description: Exact, deterministically converted maintenance measurement recorded.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Measurement"}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/executions/{executionID}/observations:
    post:
      summary: Record an observation
      tags: [Executions]
      operationId: recordObservation
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ExecutionID"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "201": {description: "Human, voice, vision-candidate, or external observation recorded."}
  /api/v1/executions/{executionID}/actions:
    post:
      summary: Record a maintenance action
      tags: [Executions]
      operationId: recordMaintenanceAction
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ExecutionID"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "201": {description: Non-control maintenance action evidence recorded.}
  /api/v1/executions/{executionID}/decisions:
    post:
      summary: Record a technician decision
      tags: [Executions]
      operationId: recordExecutionDecision
      description: Captures a semantic decision, rationale, alternatives, and context rather than UI coordinates.
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ExecutionID"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [client_event_id, decision, rationale, decided_at]
              properties:
                client_event_id: {type: string, format: uuid}
                device_id: {type: string}
                created_at_device: {type: string, format: date-time}
                step_id: {type: string, format: uuid}
                component_id: {type: string, format: uuid}
                decision: {type: string, minLength: 1}
                rationale: {type: string, minLength: 1}
                alternatives: {type: array, items: {type: string}}
                decided_at: {type: string, format: date-time}
      responses:
        "201": {description: Semantic technician decision recorded.}
        "400": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/executions/{executionID}/complete:
    post:
      summary: Complete an execution
      tags: [Executions]
      operationId: completeExecution
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/ExecutionID"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "200": {description: Fully completed execution closed with an outcome summary.}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: skawld_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token for native/mobile Authorization Code + PKCE.
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
...
    ExecutionID:
      in: path
      name: executionID
      required: true
      schema: {type: string, format: uuid}
...
  requestBodies:
    ExpectedVersion:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [expected_version]
            properties:
              expected_version: {type: integer, format: int64, minimum: 1}
  responses:
    Problem:
      description: RFC 9457-style problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Execution:
      type: object
      required: [id, organization_id, site_id, asset_id, purpose, state, version, steps, prerequisites, measurements, observations, actions, decisions, updated_at]
      properties:
        id: {type: string, format: uuid}
        organization_id: {type: string, format: uuid}
        site_id: {type: string, format: uuid}
        incident_id: {type: string, format: uuid}
        asset_id: {type: string, format: uuid}
        asset_tag: {type: string}
        purpose: {type: string}
        state: {type: string, enum: [ASSIGNED, IN_PROGRESS, COMPLETED]}
        version: {type: integer, format: int64}
        updated_at: {type: string, format: date-time}
        steps:
          type: array
          items: {$ref: "#/components/schemas/ExecutionStep"}
        prerequisites: {type: array, items: {type: object}}
        measurements:
          type: array
          items: {$ref: "#/components/schemas/Measurement"}
        observations: {type: array, items: {type: object}}
        actions: {type: array, items: {type: object}}
        decisions: {type: array, items: {type: object}}
    ExecutionStep:
      type: object
      required: [id, key, sequence, title, state, risk_level, version]
      properties:
        id: {type: string, format: uuid}
        key: {type: string}
        sequence: {type: integer}
        title: {type: string}
        state: {type: string, enum: [PENDING, COMPLETED, BLOCKED]}
        risk_level:
          type: string
          enum: [INFORMATIONAL, ADVISORY, OPERATIONAL_LOW, SAFETY_SIGNIFICANT]
        required_prerequisite: {type: string}
        blocked_reason: {type: string}
        version: {type: integer, format: int64}
    PrerequisiteVerificationCommand:
      type: object
      additionalProperties: false
      required: [type, status, external_reference, verified_at]
      properties:
        type:
          type: string
          enum: [WORK_PERMIT, ENERGY_ISOLATION, GAS_TEST, SUPERVISOR, COMPETENCY]
        status: {type: string, enum: [VERIFIED, REJECTED, EXPIRED]}
        external_reference: {type: string, minLength: 1}
        verified_at: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
    MeasurementCommand:
      type: object
      additionalProperties: false
      required: [client_event_id, measurement_type, value, unit, source, data_quality, verification_status, observed_at]
      properties:
        client_event_id: {type: string, format: uuid}
        device_id: {type: string}
        created_at_device: {type: string, format: date-time}
        component_id: {type: string, format: uuid}
        measurement_type:
          type: string
          enum: [VIBRATION_VELOCITY, TEMPERATURE]
        value:
          type: string
          pattern: '^-?(0|[1-9][0-9]*)(\.[0-9]{1,6})?$'
        unit: {type: string, enum: [MM_PER_S, IN_PER_S, DEG_C, DEG_F]}
        source: {type: string, enum: [MANUAL, INSTRUMENT, EXTERNAL_SYSTEM]}
        data_quality: {type: string, enum: [GOOD, QUESTIONABLE, BAD, UNKNOWN]}
        verification_status: {type: string, enum: [UNVERIFIED, VERIFIED, REJECTED]}
        observed_at: {type: string, format: date-time}
    Measurement:
      type: object
      properties:
        id: {type: string, format: uuid}
        measurement_type: {type: string}
        value: {type: string}
        unit: {type: string}
        original_value: {type: string}
        original_unit: {type: string}
        source: {type: string}
        data_quality: {type: string}
        verification_status: {type: string}
        observed_at: {type: string, format: date-time}
        received_at_server: {type: string, format: date-time}
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string